ci: :bug: Order, Common/Tape initial bug fixes
