+++ src/pages/Common/Coil/Log/index.jsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import ProductionLog from './ProductionLog';
 import RMOrderAgainstLog from './RMOrderAgainstLog';
-import RMTransferLog from './RMTransferLog/RMTransferLog';
-import SFGTransferLog from './SFGTransferLog/SFGTransferLog';
+import RMTransferLog from './RMTransfer';
+import SFGTransferLog from './SFGTransfer';
 
 export default function Index() {
 	useEffect(() => {
@@ -11,12 +11,12 @@ export default function Index() {
 	return (
 		<div>
 			<SFGTransferLog />
-			<hr className='border-secondary-content my-6 border-2 border-dashed' />
-			{/* <RMTransferLog />
+			<hr className='my-6 border-2 border-dashed border-secondary-content' />
+			<RMTransferLog />
 			<hr className='my-6 border-2 border-dashed border-secondary-content' />
 			<ProductionLog />
 			<hr className='my-6 border-2 border-dashed border-secondary-content' />
-			<RMOrderAgainstLog /> */}
+			<RMOrderAgainstLog />
 		</div>
 	);
 }
