+++ src/pages/Common/Tape/Log/RMOrderAgainstLog/index.jsx
@@ -1,20 +1,17 @@
 import { Suspense } from '@/components/Feedback';
 import { DeleteModal } from '@/components/Modal';
 import ReactTable from '@/components/Table';
-import { useAccess, useFetchFunc } from '@/hooks';
-import {
-	useCommonOrderAgainstTapeRMLog,
-	useCommonTapeRMLog,
-} from '@/state/Common';
+import { useAccess } from '@/hooks';
+import { useCommonOrderAgainstTapeRMLog } from '@/state/Common';
 import { DateTime, EditDelete } from '@/ui';
 import PageInfo from '@/util/PageInfo';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 
 import {
 	useMaterialInfo,
 	useMaterialTrxAgainstOrderDescription,
 } from '@/state/Store';
-import { SFG_TRX_NULL } from '@/util/Schema';
+
 import RMAddOrUpdate from './AddOrUpdate';
 
 export default function Index() {
@@ -25,11 +22,7 @@ export default function Index() {
 	const { invalidateQuery: invalidateMaterialTrx } =
 		useMaterialTrxAgainstOrderDescription();
 
-	const info = new PageInfo(
-		'RM Order Against Tape Log',
-		url,
-		'common__tape_log'
-	);
+	const info = new PageInfo('Store/Stock -> Tape', url, 'common__tape_log');
 	const haveAccess = useAccess(info.getTab());
 
 	const columns = useMemo(
@@ -83,7 +76,9 @@ export default function Index() {
 				header: 'Used QTY',
 				enableColumnFilter: false,
 				cell: (info) => (
-					<span className='capitalize'>{info.getValue()}</span>
+					<span className='capitalize'>
+						{Number(info.getValue())}
+					</span>
 				),
 			},
 			{
@@ -94,16 +89,11 @@ export default function Index() {
 			},
 			{
 				accessorKey: 'created_by_name',
-				header: 'Issued By',
+				header: 'Created By',
 				enableColumnFilter: false,
 				cell: (info) => info.getValue(),
 			},
-			{
-				accessorKey: 'remarks',
-				header: 'Remarks',
-				enableColumnFilter: false,
-				cell: (info) => info.getValue(),
-			},
+
 			{
 				accessorKey: 'created_at',
 				header: 'Created',
@@ -122,6 +112,12 @@ export default function Index() {
 				cell: (info) => {
 					return <DateTime date={info.getValue()} />;
 				},
+			},
+			{
+				accessorKey: 'remarks',
+				header: 'Remarks',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
 			},
 			{
 				accessorKey: 'actions',
