+++ src/pages/Vislon/TeethMolding/Log/RMOrderAgainstLog/AddOrUpdate.jsx
@@ -1,5 +1,4 @@
 import { AddModal } from '@/components/Modal';
-import { useAuth } from '@/context/auth';
 import { useFetchForRhfReset, useRHF, useUpdateFunc } from '@/hooks';
 import { useCommonMaterialTrx } from '@/state/Common';
 import {
